Revert "Undo reorganization for cleaner diff (REVERT ME)"
This reverts commit 5b76ccda93abbdeb922c80d5d271cc84cdae3e21.
